+++ examples/calculator.rs
@@ -34,8 +34,6 @@ fn main() {
     let mut out = Output::new(border, border, 360, 140, "");
     out.set_text_size(30);
     out.set_value("0");
-    
-    let out = Rc::from(RefCell::from(out));
 
     let mut but_ce = Button::new(column1, row1, but_width, but_height, "CE");
     let mut but_c = Button::new(column2, row1, but_width, but_height, "C");
@@ -94,14 +92,14 @@ fn main() {
 
     for but in but_vec {
         but.set_color(Color::Light2);
-        let out = out.clone();
+        let mut out = out.clone();
         let mut txt = txt.clone();
         let op = op.clone();
         let label = but.label();
         app::set_callback(
             but,
             Box::new(move || {
-                if out.borrow().value() == "0" {
+                if out.value() == "0" {
                     txt.borrow_mut().clear();
                 }
                 if op.borrow().as_str() == "=" {
@@ -110,7 +108,7 @@ fn main() {
                     op.borrow_mut().push_str("_");
                 }
                 txt.borrow_mut().push_str(&label);
-                out.borrow_mut().set_value(txt.borrow().as_str());
+                out.set_value(txt.borrow().as_str());
             }),
         );
     }
@@ -125,60 +123,60 @@ fn main() {
             &mut but,
             Box::new(move || {
                 old_val.borrow_mut().clear();
-                old_val.borrow_mut().push_str(&out.borrow().value());
+                old_val.borrow_mut().push_str(&out.value());
                 op.borrow_mut().clear();
                 op.borrow_mut().push_str(&label);
                 txt.borrow_mut().clear();
                 txt.borrow_mut().push('0');
-                out.borrow_mut().set_value(txt.borrow().as_str());
+                out.set_value(txt.borrow().as_str());
             }),
         );
     }
 
     let mut txt_cloned = txt.clone();
     let old_val_cloned = old_val.clone();
-    let out_c = out.clone();
+    let mut out_c = out.clone();
     app::set_callback(
         & mut but_ce,
         Box::new(move || {
             txt_cloned.borrow_mut().clear();
             old_val_cloned.borrow_mut().clear();
             txt_cloned.borrow_mut().push('0');
-            out_c.borrow_mut().set_value(txt_cloned.borrow().as_str());
+            out_c.set_value(txt_cloned.borrow().as_str());
         }),
     );
 
     let mut txt_cloned = txt.clone();
-    let out_c = out.clone();
+    let mut out_c = out.clone();
     app::set_callback(
         & mut but_c,
         Box::new(move || {
             txt_cloned.borrow_mut().clear();
             txt_cloned.borrow_mut().push('0');
-            out_c.borrow_mut().set_value(txt_cloned.borrow().as_str());
+            out_c.set_value(txt_cloned.borrow().as_str());
         }),
     );
 
     let mut txt_cloned = txt.clone();
-    let out_c = out.clone();
+    let mut out_c = out.clone();
     app::set_callback(
         & mut but_back,
         Box::new(move || {
-            let val = out_c.borrow().value();
+            let val = out_c.value();
             if val.len() > 1 {
                 txt_cloned.borrow_mut().pop();
-                out_c.borrow_mut().set_value(txt_cloned.borrow().as_str());
+                out_c.set_value(txt_cloned.borrow().as_str());
             }
         }),
     );
 
     let mut txt_cloned = txt.clone();
     let op_cloned = op.clone();
-    let out_c = out.clone();
+    let mut out_c = out.clone();
     app::set_callback(
         & mut but_eq,
         Box::new(move || {
-            new_val = out_c.borrow().value();
+            new_val = out_c.value();
             let old: f64 = old_val.borrow().parse().unwrap();
             let new: f64 = new_val.parse().unwrap();
             let val = match op_cloned.borrow().as_str() {
@@ -192,11 +190,11 @@ fn main() {
             op_cloned.borrow_mut().push_str("_");
             txt_cloned.borrow_mut().clear();
             txt_cloned.borrow_mut().push_str(val.to_string().as_str());
-            out_c.borrow_mut().set_value(txt_cloned.borrow().as_str());
+            out_c.set_value(txt_cloned.borrow().as_str());
         }),
     );
 
-    let out_c = out.clone();
+    let mut out_c = out.clone();
     app::set_callback(
         & mut but_dot,
         Box::new(move || {
@@ -204,12 +202,12 @@ fn main() {
                 txt.borrow_mut().clear();
                 op.borrow_mut().clear();
                 op.borrow_mut().push_str("_");
-                out_c.borrow_mut().set_value("0.");
+                out_c.set_value("0.");
                 txt.borrow_mut().push_str("0.");
             }
             if !txt.borrow().contains(".") {
                 txt.borrow_mut().push_str(".");
-                out.borrow_mut().set_value(txt.borrow().as_str());
+                out.set_value(txt.borrow().as_str());
             }
         }),
     );