+++ examples/threads.rs
@@ -1,6 +1,5 @@
 use fltk::{app::*, button::*, frame::*, window::*};
 use std::{thread, time};
-use std::sync::{Arc, Mutex};
 
 fn main() {
     let app = App::default();
@@ -8,18 +7,16 @@ fn main() {
         .with_size(400, 300)
         .center_screen()
         .with_label("Hello");
+    let frame = Frame::new(0, 0, 400, 200, "");
     let mut but = Button::new(160, 210, 80, 40, "Click me!");
-
-    let frame = Arc::from(Mutex::from(Frame::new(0, 0, 400, 200, "")));
 
     wind.show();
 
     but.set_callback(Box::new(move || {
-        let frame = frame.clone();
+        let mut frame = frame.clone();
         thread::spawn(move|| {
             for i in 0..1000 {
                 thread::sleep(time::Duration::from_millis(10));
-                let mut frame = frame.lock().unwrap();
                 frame.set_label(format!("Hello {}", i).as_str());
             }
         });
@@ -34,20 +31,19 @@ fn main() {
 //         .with_size(400, 300)
 //         .center_screen()
 //         .with_label("Hello");
-//     let frame = Frame::new(0, 0, 400, 200, "");
 //     let mut but = Button::new(160, 210, 80, 40, "Click me!");
-//
+//     use std::sync::{Arc, Mutex};
+//     let frame = Arc::from(Mutex::from(Frame::new(0, 0, 400, 200, "")));
 //     wind.show();
-//
 //     but.set_callback(Box::new(move || {
-//         let mut frame = unsafe { frame.memcpy() };
+//         let frame = frame.clone();
 //         thread::spawn(move|| {
 //             for i in 0..1000 {
 //                 thread::sleep(time::Duration::from_millis(10));
+//                 let mut frame = frame.lock().unwrap();
 //                 frame.set_label(format!("Hello {}", i).as_str());
 //             }
 //         });
 //     }));
-//
 //     app.run().unwrap();
 // }